test: split qbft aggregator rejection test
